# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export type AppNotificationType = typeof AppNotificationType[keyof typeof AppNotificationType];


export const AppNotificationType = {
  info: 'info',
  success: 'success',
  warning: 'warning',
  promo: 'promo',
} as const;

export interface AppNotification {
  id: string;
  title: string;
  message: string;
  type: AppNotificationType;
  startAt: number;
  endAt: number;
}

export type AdminStateConfig = {[key: string]: unknown};

export interface AdminState {
  config: AdminStateConfig;
  notifications: AppNotification[];
  users: unknown[];
  deposits: unknown[];
  withdrawals: unknown[];
  tokenPackages: unknown[];
  inventory: unknown[];
  socialTasks: unknown[];
  promoCodes: unknown[];
  broadcasts: unknown[];
  tickets: unknown[];
  products: unknown[];
  referrers: unknown[];
}

/**
 * Allowlisted config keys (settings, game_overrides, ticket_overrides, referral_config, daily_checkin)
 */
export type RuntimeConfigConfig = {[key: string]: unknown};

/**
 * Mini-app safe public runtime config. Never contains PII or sensitive admin data.
 */
export interface RuntimeConfig {
  /** Allowlisted config keys (settings, game_overrides, ticket_overrides, referral_config, daily_checkin) */
  config: RuntimeConfigConfig;
  notifications: AppNotification[];
}