# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/state:
    get:
      operationId: getAdminState
      tags: [admin]
      summary: Get full admin state (auth required)
      description: Returns all configuration and entity lists. Requires an active admin session (JWT cookie). Use /admin/runtime-config for unauthenticated mini-app access.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Full admin state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminState"
        "401":
          description: Unauthorized
  /admin/config/{key}:
    put:
      operationId: putAdminConfig
      tags: [admin]
      summary: Update a config section
      security:
        - cookieAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
            enum:
              - settings
              - game_overrides
              - ticket_overrides
              - cms
              - finance
              - security
              - backup
              - referral_config
              - daily_checkin
              - api_keys
              - roles
      requestBody:
        required: true
        content:
          application/json:
            schema: {}
      responses:
        "200":
          description: Config updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: skz_admin_token
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    AppNotification:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, success, warning, promo]
        startAt:
          type: number
        endAt:
          type: number
      required:
        - id
        - title
        - message
        - type
        - startAt
        - endAt
    AdminState:
      type: object
      properties:
        config:
          type: object
          additionalProperties: {}
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/AppNotification"
        users:
          type: array
          items: {}
        deposits:
          type: array
          items: {}
        withdrawals:
          type: array
          items: {}
        tokenPackages:
          type: array
          items: {}
        inventory:
          type: array
          items: {}
        socialTasks:
          type: array
          items: {}
        promoCodes:
          type: array
          items: {}
        broadcasts:
          type: array
          items: {}
        tickets:
          type: array
          items: {}
        products:
          type: array
          items: {}
        referrers:
          type: array
          items: {}
      required:
        - config
        - notifications
        - users
        - deposits
        - withdrawals
        - tokenPackages
        - inventory
        - socialTasks
        - promoCodes
        - broadcasts
        - tickets
        - products
        - referrers
...