# lib/api-zod/src/generated/types/adminState.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminStateConfig } from './adminStateConfig';
import type { AppNotification } from './appNotification';

export interface AdminState {
  config: AdminStateConfig;
  notifications: AppNotification[];
  users: unknown[];
  deposits: unknown[];
  withdrawals: unknown[];
  tokenPackages: unknown[];
  inventory: unknown[];
  socialTasks: unknown[];
  promoCodes: unknown[];
  broadcasts: unknown[];
  tickets: unknown[];
  products: unknown[];
  referrers: unknown[];
}