# lib/api-zod/src/generated/types/appNotification.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppNotificationType } from './appNotificationType';

export interface AppNotification {
  id: string;
  title: string;
  message: string;
  type: AppNotificationType;
  startAt: number;
  endAt: number;
}